# docs/swagger.yaml
basePath: /api
definitions:
...
  controllers.RentalRequest:
    properties:
      end_date:
        description: 'Format: YYYY-MM-DD'
        example: "2025-01-05"
        type: string
      machine_id:
        example: uuid-string
        type: string
      start_date:
        description: 'Format: YYYY-MM-DD'
        example: "2025-01-01"
        type: string
    type: object
  controllers.RentalStatusUpdate:
    properties:
      status:
        description: approved, rejected, completed
        example: approved
        type: string
    type: object
...
  models.Machine:
    properties:
      category:
        description: New Fields for Search
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      listing_type:
        description: sale, rent, both
        type: string
      location:
        description: e.g., "Faridabad, Haryana"
        type: string
      manufacturer:
        type: string
      model_number:
        type: string
      price_for_sale:
        type: number
      rental_price_per_day:
        type: number
      rental_price_per_month:
        type: number
      security_deposit:
        type: number
      seller_id:
        type: integer
      specs:
        description: 'UPDATED: Added swaggertype:"object" to fix Swagger generation'
        type: object
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      year_of_manufacture:
        type: integer
    type: object
...
  models.Rental:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      machine:
        allOf:
        - $ref: '#/definitions/models.Machine'
        description: Preloads (Optional, for returning full details)
      machine_id:
        type: string
      platform_fee:
        description: e.g. 5%
        type: number
      renter_id:
        description: Matches Auth Service ID type
        type: integer
      security_deposit:
        type: number
      start_date:
        type: string
      status:
        description: 'Status Flow: pending -> approved -> active -> completed (or
          rejected/cancelled)'
        type: string
      total_amount:
        description: Financials
        type: number
      updated_at:
        type: string
    type: object
host: localhost:1324
info:
  contact: {}
  description: Backend API for Vishwakarma Setu B2B Marketplace.
  title: Vishwakarma Setu API
  version: "1.0"
paths:
...
  /rentals:
    post:
      consumes:
      - application/json
      description: Initiate a rental request. Calculates fees based on duration. Requires
        Renter Auth.
      parameters:
      - description: Rental Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.RentalRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Rental'
        "400":
          description: Invalid input or dates
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Machine not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Request to rent a machine
      tags:
      - Rentals
  /rentals/{id}/status:
    put:
      consumes:
      - application/json
      description: Approve, reject, or complete a rental. Only the machine owner can
        do this.
      parameters:
      - description: Rental ID
        in: path
        name: id
        required: true
        type: string
      - description: New Status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/controllers.RentalStatusUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Rental'
        "403":
          description: Not authorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update rental status
      tags:
      - Rentals
  /rentals/manage:
    get:
      description: Retrieve all incoming rental requests for machines owned by the
        logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Rental'
            type: array
      security:
      - BearerAuth: []
      summary: Get rental requests for my machines
      tags:
      - Rentals
  /rentals/my:
    get:
      description: Retrieve all rental requests made by the logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Rental'
            type: array
      security:
      - BearerAuth: []
      summary: Get my rental history
      tags:
      - Rentals
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"